+++ src/lib.rs
@@ -1,18 +1,3 @@
 #![allow(clippy::toplevel_ref_arg)]
 pub mod routes;
-use crate::routes::{health_check, subscribe};
-use actix_web::dev::Server;
-use actix_web::{web, App, HttpServer};
-use std::net::TcpListener;
-
-pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
-    let server = HttpServer::new(|| {
-        App::new()
-            .route("/health_check", web::get().to(health_check))
-            .route("/subscriptions", web::post().to(subscribe))
-    })
-    .listen(listener)?
-    .run();
-
-    Ok(server)
-}
+pub mod startup;
