+++ tests/health_check.rs
@@ -1,4 +1,5 @@
 use std::net::TcpListener;
+use zero2prod::startup::run;
 
 #[actix_web::test]
 async fn health_check_works() {
@@ -73,7 +74,7 @@ fn spawn_app() -> String {
     let fail_text = "Failed to bind random port";
     let listener = TcpListener::bind("127.0.0.1:0").expect(fail_text);
     let port = listener.local_addr().unwrap().port();
-    let server = zero2prod::run(listener).expect("Failed to bind address");
+    let server = run(listener).expect("Failed to bind address");
     let _ = actix_web::rt::spawn(server);
     format!("http://127.0.0.1:{}", port)
 }
